+++ app/src/main/java/coolpharaoh/tee/speicher/tea/timer/views/show_tea/ShowTeaViewModel.java
@@ -73,7 +73,7 @@ class ShowTeaViewModel {
         }
     }
 
-    int getAmount() {
+    double getAmount() {
         return teaRepository.getTeaById(teaId).getAmount();
     }
